Fix poster loading on newer Roku players
Also fix the if/else block in MovieObject which was completely broken
(no idea what I was thinking there).

It ends up that on my Roku 2 if you set the HDPosterUrl but aren't
actually using it (IsHD is false), it causes the SD poster loading to
only load one of them. Since I've never actually implemented HD support
I've disabled this block of code for now.