Check the server for Setup-SD.png
There are some situations where HMS could get stuck and not allow you to
edit the URL of the server. eg. enter an IP address of a http server
without a directory listing.

This commits requires that the server have /Setup-SD.png available. If
it isn't it will fall into the URL editing dialog.