Switch back to using roGridScreen
roPosterScreen loads quickly, but with 800+ movies the selection slows
way down. roGridScreen works quickly once everything is loaded, so in
this version it only loads a few rows at a time.

When the used stops scrolling at a new row it will load that row, the
next 2 rows, and the previous row. It now also displays a Loading ...
message in the upper right while it is busy.

If the system is idle for 30 seconds it will load one unloaded row every
30s until everything has been loaded.