Setup basic control flow for SetupServerDialog
The MainScene->Init starts the SetupServerDialog, which accepts the
user's input and places it into the dialog's serverurl field. An
observer on this field runs MainScene->OnSetupServerURL when it is
changed by the dialog.

OnSetupServerURL then copies to the MainScene's serverurl field. This
triggers an observer which starts the MainLoaderTask, and sets the
serveruri field of the task.

This is just a skeleton to make sure the control flow is working right.